+++ app/src/main/java/com/example/olegweatherapp/models/onecall/ForecastOnecall.kt
@@ -22,7 +22,7 @@ data class ForecastOnecall (
 	@SerializedName("lat") val latitude : Double,
 	@SerializedName("lon") val longitude : Double,
 	@SerializedName("timezone") val timezone : String,
-	@SerializedName("timezone_offset") val timezoneOffset : Int,
+	@SerializedName("timezone_offset") val timezoneOffset : Double,
 	@SerializedName("current") val current : Current,
 	@SerializedName("minutely") val minutely : List<Minutely>,
 	@SerializedName("hourly") val hourly : List<Hourly>,
